+++ packages/generators/CHANGELOG.md
@@ -1,4 +1,9 @@
-# Change Log
+---
+name: Generators
+menu: Changelog
+---
+
+# @jobber/generators: Change Log
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.