+++ packages/generators/CHANGELOG.md
@@ -8,6 +8,14 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/GetJobber/atlantis/compare/@jobber/generators@0.4.0...@jobber/generators@0.4.1) (2021-03-03)
+
+**Note:** Version bump only for package @jobber/generators
+
+
+
+
+
 # [0.4.0](https://github.com/GetJobber/atlantis/compare/@jobber/generators@0.1.0...@jobber/generators@0.4.0) (2020-04-08)